Deleted studentID collection from membership access requests, modified 'position' to 'role' for insertion. Added RLS to Members table and modified API naming for adding new members.
