+++ src/app/api/members/[id]/route.ts
@@ -26,36 +26,3 @@ export async function GET(request: NextRequest) {
   console.log("Member Best Single Results: ", data);
   return NextResponse.json(data, { status: 200 });
 }
-
-export async function POST(request: NextRequest) {
-  const { id, name, email, faculty, birthdate, wca_id } =
-    await request.json();
-  if (!name || !faculty) {
-    console.log("missing: ", { id, name, faculty });
-    return NextResponse.json(
-      { error: "Missing required fields" },
-      { status: 400 }
-    );
-  }
-  const { data, error } = await supabase
-    .from("Members")
-    .insert([
-      {
-        name: name,
-        email: email,
-        membership: true,
-        position: null,
-        faculty: faculty,
-        birthdate: birthdate,
-        wca_id: wca_id,
-      },
-    ])
-    .select()
-    .single();
-  if (error) {
-    console.error("Error inserting member: ", error.message);
-    return NextResponse.json({ error: error.message }, { status: 500 });
-  }
-  console.log("Inserted member: ", data);   
-  return NextResponse.json(data, { status: 201 });
-}
